backend/requirements.txt
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
boto3==1.42.63
botocore==1.42.63
certifi==2026.2.25
cffi==2.0.0
clerk-backend-api==4.2.0
click==8.1.8
cryptography==45.0.7
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.128.8
greenlet==3.2.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.0.2
pgvector==0.4.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.49.3
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.39.0
